ENH: CaptureHandler supports pptx. Fixes #161

+++ gramex/apps/capture/chromecapture.js
@@ -2,13 +2,16 @@
 /* eslint-disable no-console */
 
 const puppeteer = require('puppeteer')
+const image_size = require('fast-image-size')
 const bodyParser = require('body-parser')
 const minimist = require('minimist')
 const express = require('express')
 const cookie = require('cookie')
 const path = require('path')
+const url = require('url')
 const tmp = require('tmp')
 const fs = require('fs')
+const _ = require('lodash')
 
 const default_port = 8090
 const version = '1.0.0'
@@ -19,10 +22,41 @@ const homepage = path.join(folder, 'index.html')
 let browser, page, app, server
 let render_dir = folder             // Used by render() to save the file
 
+const pptx_size = {
+  'A3':     [14,   10.5  ],
+  'A4':     [10.83, 7.5  ],
+  'Letter': [11,    8.5  ],
+  '16x10':  [10,    6.25 ],
+  '16x9':   [10,    5.625],
+  '4x3':    [10,    7.5  ]
+}
 
 function delay(ms) {
   return new Promise(res => setTimeout(res, ms))
 }
+
+
+async function screenshot(page, options, selector) {
+  // If a previous clip was set, remove it
+  delete options.clip
+  // Apply clip based on selector
+  if (selector) {
+    const rect = await page.evaluate(
+      function(selector) {
+        let el = document.querySelector(selector)
+        if (!el) return
+        let rect = el.getBoundingClientRect()
+        return {x: rect.x, y: rect.y, width: rect.width, height: rect.height}
+      },
+      selector)
+    if (!rect)
+      throw new Error('No selector ' + selector)
+    options.clip = rect
+  }
+  // Take the screenshot
+  await page.screenshot(options)
+}
+
 
 async function render(q) {
   console.log('Opening', q.url)
@@ -65,46 +99,74 @@ async function render(q) {
       printBackground: true
     })
   } else {
-    await page.setViewport({
+    const viewport = {
       width: +q.width || 1200,
       height: +q.height || 768,
       deviceScaleFactor: +q.scale || 1
-    })
+    }
+    await page.setViewport(viewport)
     const options = {
       path: target,
       fullPage: !q.height && !q.selector  // If height and selector not specified, use full height
     }
-    if (q.selector) {
-      const rect = await page.evaluate(
-        function(selector) {
-          let el = document.querySelector(selector)
-          if (!el) return
-          let rect = el.getBoundingClientRect()
-          return {x: rect.x, y: rect.y, width: rect.width, height: rect.height}
-        },
-        q.selector)
-      if (!rect)
-        throw new Error('No selector ' + q.selector)
-      options.clip = rect
+    if (ext == 'pptx') {
+      const officegen = require('officegen')
+      const pptx = officegen('pptx')
+      // Convert to arrays
+      for (let key of ['selector', 'title', 'x', 'y'])
+        if (!Array.isArray(q[key]))
+          q[key] = [q[key]]
+      const image_files = []
+      for (const [index, [selector, title, x, y]] of _.zip(q.selector, q.title, q.x, q.y).entries()) {
+        options.path = target.replace(/\.pptx$/, '.' + index + '.png')
+        image_files.push(options.path)
+        await screenshot(page, options, selector)
+        const fmt = pptx_size[q.layout in pptx_size ? q.layout : '4x3']
+        pptx.setSlideSize(fmt[0] * 72, fmt[1] * 72)
+        const slide = pptx.makeNewSlide()
+        const dpi = +(q.dpi || 96)
+        const size = image_size(options.path)
+        slide.addImage(options.path, {
+          x: typeof x == 'undefined' ? 'c' : x / dpi * 72,
+          y: typeof y == 'undefined' ? 'c' : y / dpi * 72,
+          cx: size.width / dpi * 72,
+          cy: size.height / dpi * 72
+        })
+        if (typeof title != 'undefined')
+          slide.addText(title, 0, 0, '100%', 36)
+      }
+      await new Promise(res => {
+        const out = fs.createWriteStream(target)
+        pptx.generate(out)
+        out.on('close', res)
+      })
+      image_files.forEach(path => fs.unlinkSync(path))
+    } else {
+      await screenshot(page, options, q.selector)
     }
-    await page.screenshot(options)
   }
   return {path: target, file: file}
 }
 
 function webapp(req, res) {
-  var q = Object.assign({}, req.query, req.body)
+  let q = Object.assign({}, req.query, req.body)
   if (!q.url)
     return res.sendFile(homepage)
   q.cookie = q.cookie || req.headers.cookie
   render(q)
     .then((info) => {
-      res.setHeader('Content-Disposition', 'attachment; filename=' + info.file)
-      res.sendFile(info.path, (err) => {
-        if (err)
-          console.error('Error sending file', err)
-        fs.unlinkSync(info.path)
-      })
+      if (fs.existsSync(info.path)) {
+        res.setHeader('Content-Disposition', 'attachment; filename=' + info.file)
+        res.sendFile(info.path, (err) => {
+          if (err)
+            console.error('Error sending file', err)
+          fs.unlinkSync(info.path)
+        })
+      } else {
+        console.error('Missing file', info.path)
+        res.setHeader('Content-Type', 'text/plain')
+        res.send('Missing output file.')
+      }
     })
     .catch((e) => {
       res.setHeader('Content-Type', 'text/plain')
